// ie421_hft_spring_2023_group_04/src/netcap_app/src/types/RawPCAPTypes.ts
export type RawPCAPTableType = {
    id: number;
    time: number;
    length: number;
    source_ip: string;
    destination_ip: string;
    network_protocol: string;
    ip_protocol: string;
    tcp_flags: string;
    application_protocol: string;
};

export type PCAPRawDetailsType = {
    rawhex: string;
    sll_packet_type: string;
    sll_address_type: string;
    sll_address_length: string;
    sll_source_address: string;
    sll_protocol: string;
    ip_version?: string;
    ip_header_length?: string;
    ip_type_of_service?: string;
    ip_total_length?: string;
    ip_id?: string;
    ip_offset?: string;
    ip_time_to_live?: string;
    ip_protocol?: string;
    ip_checksum?: string;
    ip_src_ip?: string;
    ip_dst_ip?: string;
    tcp_src_port?: string;
    tcp_dst_port?: string;
    tcp_sequence_num?: string;
    tcp_ack_number?: string;
    tcp_offset?: string;
    tcp_flags?: string;
    tcp_window?: string;
    tcp_checksum?: string;
    tcp_urgent_pointer?: string;
    udp_src_port?: string;
    udp_dst_port?: string;
    udp_total_length?: string;
    udp_checksum?: string;
    fix_begin_string?: string;
    fix_body_length?: string;
    fix_msg_type?: string;
    fix_sender_comp_id?: string;
    fix_target_comp_id?: string;
    fix_msg_seq_num?: string;
    fix_sent_tstamp?: string;
    fix_cl0rdid?: string;
    fix_symbol?: string;
    fix_ord_status?: string;
    fix_side?: string;
    fix_ord_type?: string;
    fix_ord_qty?: string;
    fix_time_in_force?: string;
    fix_price?: string;
    fix_mdreq_id?: string;
    fix_checksum?: string;
};